backend-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.1
httpx==0.26.0
python-dotenv==1.0.0
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
email-validator==2.1.0
slowapi==0.1.9
